+++ R/create_cost_corridor.R
@@ -31,6 +31,8 @@
 
 create_cost_corridor <- function(x, origin, destination, rescale = FALSE) { 
   
+  check_locations(x, rbind(origin, destination))
+  
   cs_rast <- terra::rast(nrow = x$nrow, ncol = x$ncol, xmin = x$extent[1], xmax = x$extent[2], ymin = x$extent[3], ymax = x$extent[4],crs = x$crs)
   
   from_coords <- get_coordinates(origin)