+++ R/create_slope_cs.R
@@ -1,14 +1,16 @@
 #' Creates a slope-based cost surface
 #' 
-#'  Creates a cost surface based on the difficulty of moving up and down slope. This function provides multiple isotropic and anisotropic cost functions that estimate the 'cost' of human movement when traversing a landscape
+#'  Creates a cost surface based on the difficulty of moving up and down slope. This function implements multiple isotropic and anisotropic cost functions that estimate the 'cost' of human movement when traversing a landscape
 #'  
-#' The supplied 'spatRaster' object must have a projected CRS
+#' The supplied 'spatRaster' object can have a projected or geographic coordinate system
 #' 
 #' @details
 #' 
-#' The following cost functions have been implemented however users may also supply their own cost functions (see examples):
+#' The following cost functions have been implemented however users may also supply their own cost function (see Examples):
 #' 
-#' "tobler", "tobler offpath", "davey", 'rees', "irmischer-clarke male", "irmischer-clarke offpath male", "irmischer-clarke female", "irmischer-clarke offpath female", "modified tobler", 'garmy', 'kondo-saino', "wheeled transport", "herzog", "llobera-sluckin", 'naismith', 'minetti', 'campbell', "campbell 2019", "sullivan"
+#' "tobler", "tobler offpath", "davey", 'rees', "irmischer-clarke male", "irmischer-clarke offpath male", "irmischer-clarke female", "irmischer-clarke offpath female", "modified tobler", 'garmy', 'kondo-saino', "wheeled transport", "herzog", "llobera-sluckin", "naismith", "minetti", "campbell","campbell 2019 1","campbell 2019 5" ,"campbell 2019 10","campbell 2019 15","campbell 2019 20","campbell 2019 25","campbell 2019 30","campbell 2019 35","campbell 2019 40","campbell 2019 45","campbell 2019 50","campbell 2019 55","campbell 2019 60","campbell 2019 65","campbell 2019 70","campbell 2019 75","campbell 2019 80","campbell 2019 85","campbell 2019 90","campbell 2019 95","campbell 2019 99", "sullivan 167","sullivan 5", "sullivan 833"
+#' 
+#' Multiple travel rate percentiles implemented for campbell 2019 and sullivan, e.g. "campbell 2019 50" is the 50th percentile
 #'
 #' @param x \code{SpatRaster} Digital Elevation Model (DEM)
 #'
@@ -19,8 +21,6 @@
 #' @param crit_slope \code{numeric} value. Critical Slope (in percentage) is 'the transition where switchbacks become more effective than direct uphill or downhill paths'. Cost of climbing the critical slope is twice as high as those for moving on flat terrain and is used for estimating the cost of using wheeled vehicles. Default value is 12, which is the postulated maximum gradient traversable by ancient transport (Verhagen and Jeneson, 2012). Critical slope only used in 'wheeled transport' cost function
 #'
 #' @param max_slope \code{numeric} value. Maximum percentage slope that is traversable. Slope values that are greater than the specified max_slope are given a conductivity value of 0. If cost_function argument is 'campbell 2019' or 'campbell' then max_slope is fixed at 30 degrees slope to reflect the maximum slope that the cost function is parametised to. NULL (default)
-#'
-#' @param percentile \code{numeric} value. Travel rate percentile only used in 'campbell 2019' cost_function. Expected numeric values are 0.01, 0.05, 0.10, 0.15, 0.20, 0.25, 0.30, 0.35, 0.40, 0.45, 0.50, 0.55, 0.60, 0.65, 0.70, 0.75, 0.80, 0.85, 0.90, 0.95, 0.99. 0.5 (default)
 #'
 #' @param exaggeration \code{logical}. if TRUE, positive slope values (up-hill movement) multiplied by 1.99 and negative slope values (down-hill movement) multiplied by 2.31
 #'
@@ -65,15 +65,12 @@
 #' r <- terra::rast(system.file("extdata/SICILY_1000m.tif", package="leastcostpath"))
 #' 
 #' slope_cs <- create_slope_cs(x = r, cost_function = "tobler", neighbours = 4)
+#' slope_cs <- create_slope_cs(x = r, cost_function = "campbell 2019 50", neighbours = 4)
 #' slope_cs2 <- create_slope_cs(x = r, 
 #' cost_function = function(x) {(6 * exp(-3.5 * abs(x + 0.05))) / 3.6}, neighbours = 4)
 #' 
 
-create_slope_cs <- function(x, cost_function = "tobler", neighbours = 16, crit_slope = 12, max_slope = NULL, percentile = 0.5, exaggeration = FALSE) {
-  
-  if(terra::is.lonlat(x)) { 
-    stop("supplied digital elevation model (DEM) is invalid. x argument expects DEM with a projected coordinate system")
-  }
+create_slope_cs <- function(x, cost_function = "tobler", neighbours = 16, crit_slope = 12, max_slope = NULL, exaggeration = FALSE) {
     
   neighbours <- neighbourhood(neighbours = neighbours)
     
@@ -98,7 +95,7 @@ create_slope_cs <- function(x, cost_function = "tobler", neighbours = 16, crit_s
     
   ncells <- length(cells) + length(na_cells)
     
-  cf <- cost(cost_function = cost_function, crit_slope = crit_slope, percentile = percentile)
+  cf <- cost(cost_function = cost_function, crit_slope = crit_slope)
     
   if(is.function(cost_function)) { 
     message(c("Applying ", deparse(body(cost_function)[[2]]), " cost function"))
@@ -132,12 +129,11 @@ create_slope_cs <- function(x, cost_function = "tobler", neighbours = 16, crit_s
              "maxSlope" = ifelse(!is.null(max_slope), paste0(max_slope*100, "%"), NA), 
              "exaggeration" = exaggeration,
              "criticalSlope" = ifelse(test = !is.function(cost_function), yes = ifelse(test = cost_function == "wheeled transport", yes = paste0(max_slope, "%"), no = NA), no = NA),
-             "percentile" = ifelse(test = !is.function(cost_function), yes = ifelse(test = cost_function == "campbell 2019", yes = percentile, no = NA), no = NA),
              "neighbours" = sum(neighbours, na.rm = TRUE),
              "resolution" = terra::res(x), 
              "nrow" = terra::nrow(x), 
              "ncol" = terra::ncol(x), 
-             "extent" = x@ptr$extent$vector, 
+             "extent" = as.vector(terra::ext(x)), 
              "crs" = terra::crs(x, proj = TRUE))
     
   class(cs) <- "conductanceMatrix"
@@ -158,7 +154,6 @@ print.conductanceMatrix <- function(x, ...) {
     cat("\nmax slope:", x$maxSlope)
     cat("\nexaggeration:", x$exaggeration)
     cat("\ncritical slope:", x$criticalSlope)
-    cat("\npercentile:", x$percentile)
     cat("\nSpatRaster dimenions: ", x$nrow, x$ncol, prod(x$nrow, x$ncol),  "(nrow, ncol, ncell)")
     cat("\nMatrix dimensions: ", x$conductanceMatrix@Dim,  "(nrow, ncol)")
     cat("\ncrs:", x$crs)