+++ NEWS.md
@@ -1,3 +1,17 @@
+# 2.0.9 (09/04/2023)
+* removed percentile argument in create_slope_cs and replaced with their own cost_function, e.g. "campbell 2019 50"
+* Modified how extent is retrieved from a terra SpatRaster. This is now done using terra::ext()
+* calculate_distance uses Pythagorean theorem when coordinate system is projected and sf::st_distance when geographic
+* removed plot_cf()
+
+# 2.0.8 (24/03/2023)
+* create_accum_cost now allows for one or more supplied origins. Multiple accumulated cost surfaces will be summarised using a supplied function
+* exported calculate_distance, get_coordinates, and neighbourhood functions
+* Modified check_locations to stop if any locations are outside extent or not traversable. check_locations also added to functions that are suppplied locations, e.g. create_lcp
+* modified create_distance_cs to return conductanceMatrix values that correspond to the supplied spatRaster resolution
+* Added vignette
+* leastcostpath now allows for both projected and geographic coordinate systems
+
 # 2.0.7 (11/03/2023)
 * Fixed error when using terra spatVector objects in create_FETE_lcps and create_lcps
 * create_accum_cost and create_lcp_density can now take terra spatVector objects